refactor!: rename culture/server to culture/agentirc
BREAKING CHANGE: The internal Python package `culture.server` is now
`culture.agentirc`. All imports must update from `culture.server.*` to
`culture.agentirc.*`. The CLI command `culture server` and config path
`~/.culture/server.yaml` are unchanged.

AgentIRC is now the official name for the server engine in docs.
