+++ culture/agentirc/client.py
@@ -6,14 +6,14 @@ import logging
 import re
 from typing import TYPE_CHECKING
 
+from culture.agentirc.channel import Channel
+from culture.agentirc.skill import Event, EventType
 from culture.aio import maybe_await
 from culture.protocol import replies
 from culture.protocol.message import Message
-from culture.server.channel import Channel
-from culture.server.skill import Event, EventType
 
 if TYPE_CHECKING:
-    from culture.server.ircd import IRCd
+    from culture.agentirc.ircd import IRCd
 
 
 class Client:
@@ -492,7 +492,7 @@ class Client:
         )
 
     async def _send_to_client(self, target, relay, text, is_notice):
-        from culture.server.remote_client import RemoteClient
+        from culture.agentirc.remote_client import RemoteClient
 
         recipient = self.server.get_client(target)
         if not recipient:
@@ -550,7 +550,7 @@ class Client:
             await self._notify_mentions(None, text)
 
     async def _notify_mentions(self, channel_name: str | None, text: str) -> None:
-        from culture.server.remote_client import RemoteClient
+        from culture.agentirc.remote_client import RemoteClient
 
         mentioned_nicks = re.findall(r"@(\S+)", text)
         if not mentioned_nicks:
@@ -618,7 +618,7 @@ class Client:
         return flags
 
     async def _send_who_reply(self, member, channel_name: str, channel=None) -> None:
-        from culture.server.remote_client import RemoteClient  # noqa: F811
+        from culture.agentirc.remote_client import RemoteClient  # noqa: F811
 
         flags = self._build_who_flags(member, channel)
         server_name = (
@@ -660,7 +660,7 @@ class Client:
             await self.send_numeric(replies.RPL_ENDOFWHO, target, replies.MSG_ENDOFWHO)
 
     async def _handle_whois(self, msg: Message) -> None:
-        from culture.server.remote_client import RemoteClient
+        from culture.agentirc.remote_client import RemoteClient
 
         if not msg.params:
             await self.send_numeric(replies.ERR_NONICKNAMEGIVEN, "No nickname given")
