+++ culture/agentirc/rooms_util.py
