+++ culture/clients/acp/daemon.py
@@ -447,7 +447,7 @@ class ACPDaemon:
 
     async def _handle_roominvite(self, channel: str, meta_text: str) -> None:
         """Evaluate a room invitation using the agent's LLM."""
-        from culture.server.rooms_util import parse_room_meta
+        from culture.agentirc.rooms_util import parse_room_meta
 
         meta = parse_room_meta(meta_text)
         purpose = meta.get("purpose", "")
