+++ docs/use-cases/01-pair-programming.md
@@ -29,7 +29,7 @@ embedding search that complements the existing `HISTORY RECENT` and
 let agents find messages by meaning: "what did we decide about the
 embedding model?" would match messages that never contain the word
 "embedding." This is a natural extension of the `HistorySkill` that
-already lives in `culture/server/skills/history.py`.
+already lives in `culture/agentirc/skills/history.py`.
 
 The conversation starts in `#general` where Ori @mentions `spark-culture`
 with the high-level idea. The agent reads recent channel history to
@@ -63,7 +63,7 @@ class it proposes modifying, is part of the running system.
 <spark-culture> Good idea. Let me look at the current implementation
                  to see what we're working with.
 
-# Agent examines the codebase it knows: culture/server/skills/history.py.
+# Agent examines the codebase it knows: culture/agentirc/skills/history.py.
 # It sees HistorySkill with on_command dispatching to _handle_recent and
 # _handle_search, the HistoryEntry dataclass (nick, text, timestamp),
 # and the in-memory deque storage in self._channels.
@@ -205,7 +205,7 @@ class it proposes modifying, is part of the running system.
 
 1. **Ori @mentions `spark-culture` in `#general`** with a feature request. The server detects the @mention in the PRIVMSG text and sends a NOTICE to the agent's daemon.
 2. **Daemon activates** — the NOTICE triggers an agent session. The agent uses `HISTORY RECENT #general 20` to read conversation context before responding.
-3. **Agent examines real code** — it references the actual `HistorySkill` class in `culture/server/skills/history.py`, the `HistoryEntry` dataclass, the `on_command` dispatch pattern, the deque-based storage, and the `Skill` base class in `culture/server/skill.py`.
+3. **Agent examines real code** — it references the actual `HistorySkill` class in `culture/agentirc/skills/history.py`, the `HistoryEntry` dataclass, the `on_command` dispatch pattern, the deque-based storage, and the `Skill` base class in `culture/agentirc/skill.py`.
 4. **Agent proposes an architecture** grounded in the existing code — a new `_handle_semantic` method alongside `_handle_recent` and `_handle_search`, with parallel numpy vector storage per channel.
 5. **Conversation moves to DM** — Ori sends a direct PRIVMSG to the agent for a deeper design discussion. No channel noise, no overhead for other agents.
 6. **Design is refined iteratively** — embedding model selection (MiniLM for constrained Orin hardware), in-process numpy storage, lazy model loading with asyncio.Lock, graceful degradation when the model isn't installed, and the protocol extension format with similarity scores and optional threshold.
