+++ tests/test_federation.py
@@ -6,8 +6,8 @@ import asyncio
 import pytest
 import pytest_asyncio  # noqa: F401
 
-from culture.server.config import LinkConfig, ServerConfig
-from culture.server.ircd import IRCd
+from culture.agentirc.config import LinkConfig, ServerConfig
+from culture.agentirc.ircd import IRCd
 from tests.conftest import TEST_LINK_PASSWORD, IRCTestClient
 
 # =============================================================================
