+++ tests/test_room_persistence.py
@@ -8,8 +8,8 @@ import pytest
 
 def test_room_store_save_and_load():
     """Rooms can be saved to disk and loaded back."""
-    from culture.server.channel import Channel
-    from culture.server.room_store import RoomStore
+    from culture.agentirc.channel import Channel
+    from culture.agentirc.room_store import RoomStore
 
     with tempfile.TemporaryDirectory() as tmpdir:
         store = RoomStore(tmpdir)
@@ -48,8 +48,8 @@ def test_room_store_save_and_load():
 
 def test_room_store_delete():
     """Rooms can be deleted from store."""
-    from culture.server.channel import Channel
-    from culture.server.room_store import RoomStore
+    from culture.agentirc.channel import Channel
+    from culture.agentirc.room_store import RoomStore
 
     with tempfile.TemporaryDirectory() as tmpdir:
         store = RoomStore(tmpdir)
@@ -68,7 +68,7 @@ def test_room_store_delete():
 
 def test_room_store_load_empty_dir():
     """Loading from empty dir returns empty list."""
-    from culture.server.room_store import RoomStore
+    from culture.agentirc.room_store import RoomStore
 
     with tempfile.TemporaryDirectory() as tmpdir:
         store = RoomStore(tmpdir)
